Minor corrections and rearrangements
